django_ecommerce/customer_service/requirements.txt
django==3.1.12
djangorestframework==3.14.0
mysqlclient==2.1.0
requests==2.28.1